+++ tools/img-rename.js
@@ -3,7 +3,7 @@ const { renameSync } = require('fs');
 const glob = require('glob');
 
 const SOURCE = '../process/in/210108-karen+anne/AFGANG E20 rettede billeder OK/'
-const DEST   = '../dist/project-images-2/'
+const DEST   = '../dist/project-images/'
 
 // let files = glob.sync('**/*.{jpg,png,jpeg,gif}', {
 let files = glob.sync('**/*.jpg', {